# Make performance of TPCH q15 stable 