--- a/dbms/src/Common/MemoryTracker.h
+++ b/dbms/src/Common/MemoryTracker.h
@@ -111,6 +111,7 @@ namespace CurrentMemoryTracker
 {
 void disableThreshold();
 void submitLocalDeltaMemory();
+Int64 getLocalDeltaMemory();
 void alloc(Int64 size);
 void realloc(Int64 old_size, Int64 new_size);
 void free(Int64 size);